refactor(orchestrator): split capability graph modules
