+++ src/ouroboros/orchestrator/capabilities/__init__.py
@@ -7,10 +7,11 @@ from dataclasses import dataclass, field
 from enum import StrEnum
 from functools import lru_cache
 import hashlib
+import importlib
 import os
 from pathlib import Path
 import stat
-from typing import Any
+from typing import TYPE_CHECKING, Any
 import unicodedata
 
 import yaml
@@ -23,6 +24,9 @@ from ouroboros.orchestrator.mcp_tools import (
     ToolCatalogSourceMetadata,
 )
 from ouroboros.orchestrator.skill_tool_mapping import get_packaged_skill_context_keys
+
+if TYPE_CHECKING:
+    from ouroboros.orchestrator.capabilities.tool_specs import _OuroborosToolCapabilitySpec
 
 log = get_logger(__name__)
 
@@ -109,66 +113,12 @@ class CapabilityToolMetadata:
     orchestration: Mapping[str, Any] = field(default_factory=dict)
 
 
-@dataclass(frozen=True, slots=True)
-class _OuroborosToolCapabilitySpec:
-    """Explicit capability metadata spec for one Ouroboros-owned MCP tool."""
-
-    execution_mode: str
-    companions: tuple[str, ...]
-    side_effects: tuple[str, ...]
-    retry: Mapping[str, Any]
-    interrupt: Mapping[str, Any]
-    mutation_class: CapabilityMutationClass | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class LateralPersonaMetadata:
-    """Structured metadata for one lateral persona subagent lane."""
-
-    persona_id: str
-    role: str
-    prompt: Mapping[str, Any]
-    response_payload_ref: Mapping[str, Any]
-
-    def to_dict(self) -> dict[str, Any]:
-        """Serialize to JSON-safe capability metadata."""
-        return {
-            "persona_id": self.persona_id,
-            "role": self.role,
-            "prompt": dict(self.prompt),
-            "response_payload_ref": dict(self.response_payload_ref),
-        }
-
-
-@dataclass(frozen=True, slots=True)
-class LateralPersonaPanelMetadata:
-    """Structured metadata for lateral multi-persona orchestration."""
-
-    panel_id: str
-    mcp_tool: str
-    dispatch_modes: tuple[str, ...]
-    parallel_preference: str
-    sequential_fallback: Mapping[str, Any]
-    personas: tuple[LateralPersonaMetadata, ...]
-    request_model_schema: Mapping[str, Any]
-    response_payload_refs: Mapping[str, Any]
-    runtime_instruction: str
-    legacy_dispatch_modes: tuple[str, ...] = ()
-
-    def to_dict(self) -> dict[str, Any]:
-        """Serialize to JSON-safe capability metadata."""
-        return {
-            "panel_id": self.panel_id,
-            "mcp_tool": self.mcp_tool,
-            "dispatch_modes": list(self.dispatch_modes),
-            "legacy_dispatch_modes": list(self.legacy_dispatch_modes),
-            "parallel_preference": self.parallel_preference,
-            "sequential_fallback": dict(self.sequential_fallback),
-            "personas": [persona.to_dict() for persona in self.personas],
-            "request_model_schema": dict(self.request_model_schema),
-            "response_payload_refs": dict(self.response_payload_refs),
-            "runtime_instruction": self.runtime_instruction,
-        }
+_lateral_personas = importlib.import_module(f"{__name__}.lateral_personas")
+LateralPersonaMetadata = _lateral_personas.LateralPersonaMetadata
+LateralPersonaPanelMetadata = _lateral_personas.LateralPersonaPanelMetadata
+_lateral_persona_panel_metadata = _lateral_personas._lateral_persona_panel_metadata
+_lateral_persona_panel_request_schema = _lateral_personas._lateral_persona_panel_request_schema
+_pm_interview_subagent_metadata = _lateral_personas._pm_interview_subagent_metadata
 
 
 @dataclass(frozen=True, slots=True)
@@ -296,968 +246,65 @@ _DEFAULT_ATTACHED_SEMANTICS = CapabilitySemantics(
     scope=CapabilityScope.ATTACHMENT,
 )
 
-_OUROBOROS_COMPANION_FAMILIES: tuple[tuple[str, tuple[str, ...]], ...] = (
-    (
-        "auto",
-        (
-            "ouroboros_auto",
-            "ouroboros_start_auto",
-        ),
-    ),
-    (
-        "execute",
-        (
-            "ouroboros_execute_seed",
-            "ouroboros_start_execute_seed",
-            "ouroboros_cancel_execution",
-        ),
-    ),
-    (
-        "job",
-        (
-            "ouroboros_start_auto",
-            "ouroboros_start_execute_seed",
-            "ouroboros_start_evaluate",
-            "ouroboros_start_evolve_step",
-            "ouroboros_start_ralph",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-        ),
-    ),
-    (
-        "session",
-        (
-            "ouroboros_session_status",
-            "ouroboros_query_events",
-            "ouroboros_query_projection",
-            "ouroboros_ac_tree_hud",
-        ),
-    ),
-    (
-        "authoring",
-        (
-            "ouroboros_interview",
-            "ouroboros_generate_seed",
-            "ouroboros_pm_interview",
-            "ouroboros_brownfield",
-        ),
-    ),
-    (
-        "evaluation",
-        (
-            "ouroboros_evaluate",
-            "ouroboros_start_evaluate",
-            "ouroboros_checklist_verify",
-            "ouroboros_measure_drift",
-            "ouroboros_qa",
-        ),
-    ),
-    (
-        "evolution",
-        (
-            "ouroboros_evolve_step",
-            "ouroboros_start_evolve_step",
-            "ouroboros_lineage_status",
-            "ouroboros_evolve_rewind",
-            "ouroboros_ralph",
-            "ouroboros_start_ralph",
-        ),
-    ),
-    (
-        "subagent_orchestration",
-        (
-            "ouroboros_interview",
-            "ouroboros_lateral_think",
-        ),
-    ),
+_tool_specs = importlib.import_module(f"{__name__}.tool_specs")
+if not TYPE_CHECKING:
+    _OuroborosToolCapabilitySpec = _tool_specs._OuroborosToolCapabilitySpec
+_OUROBOROS_BACKGROUND_BLOCKING_COMPANIONS = _tool_specs._OUROBOROS_BACKGROUND_BLOCKING_COMPANIONS
+_OUROBOROS_BACKGROUND_LIFECYCLE_ROLE_TOOLS = _tool_specs._OUROBOROS_BACKGROUND_LIFECYCLE_ROLE_TOOLS
+_OUROBOROS_BACKGROUND_TOOLS = _tool_specs._OUROBOROS_BACKGROUND_TOOLS
+_OUROBOROS_CANCEL_METADATA = _tool_specs._OUROBOROS_CANCEL_METADATA
+_OUROBOROS_CANCEL_TOOLS = _tool_specs._OUROBOROS_CANCEL_TOOLS
+_OUROBOROS_COMPANION_FAMILIES = _tool_specs._OUROBOROS_COMPANION_FAMILIES
+_OUROBOROS_DEFAULT_EXECUTION_MODE = _tool_specs._OUROBOROS_DEFAULT_EXECUTION_MODE
+_OUROBOROS_DEFAULT_RETRY_METADATA = _tool_specs._OUROBOROS_DEFAULT_RETRY_METADATA
+_OUROBOROS_JOB_POLL_RETRY_METADATA = _tool_specs._OUROBOROS_JOB_POLL_RETRY_METADATA
+_OUROBOROS_UNSUPPORTED_RETRY_METADATA = _tool_specs._OUROBOROS_UNSUPPORTED_RETRY_METADATA
+_OUROBOROS_DEFAULT_INTERRUPT_METADATA = _tool_specs._OUROBOROS_DEFAULT_INTERRUPT_METADATA
+_OUROBOROS_BLOCKING_INTERRUPT_METADATA = _tool_specs._OUROBOROS_BLOCKING_INTERRUPT_METADATA
+_OUROBOROS_BACKGROUND_INTERRUPT_METADATA = _tool_specs._OUROBOROS_BACKGROUND_INTERRUPT_METADATA
+_OUROBOROS_TERMINAL_CONTROL_INTERRUPT_METADATA_BY_TOOL = (
+    _tool_specs._OUROBOROS_TERMINAL_CONTROL_INTERRUPT_METADATA_BY_TOOL
 )
+_OUROBOROS_UNSUPPORTED_INTERRUPT_METADATA = _tool_specs._OUROBOROS_UNSUPPORTED_INTERRUPT_METADATA
+_OUROBOROS_JOB_LIFECYCLE_SIBLING_ORDER = _tool_specs._OUROBOROS_JOB_LIFECYCLE_SIBLING_ORDER
+_OUROBOROS_MUTATION_TARGETS_BY_SIDE_EFFECT = _tool_specs._OUROBOROS_MUTATION_TARGETS_BY_SIDE_EFFECT
+_OUROBOROS_READ_ONLY_INTERRUPT_METADATA = _tool_specs._OUROBOROS_READ_ONLY_INTERRUPT_METADATA
+_OUROBOROS_SIDE_EFFECT_FREE_METADATA = _tool_specs._OUROBOROS_SIDE_EFFECT_FREE_METADATA
+_OUROBOROS_STATE_MUTATIONS_BY_TOOL = _tool_specs._OUROBOROS_STATE_MUTATIONS_BY_TOOL
+_OUROBOROS_STATUS_TOOLS = _tool_specs._OUROBOROS_STATUS_TOOLS
+_OUROBOROS_SUBAGENT_TOOLS = _tool_specs._OUROBOROS_SUBAGENT_TOOLS
+_OUROBOROS_TOOL_CAPABILITY_SPECS = _tool_specs._OUROBOROS_TOOL_CAPABILITY_SPECS
+_OUROBOROS_UNSUPPORTED_CANCEL_METADATA = _tool_specs._OUROBOROS_UNSUPPORTED_CANCEL_METADATA
+_OUROBOROS_BACKGROUND_JOB_CANCEL_METADATA = _tool_specs._OUROBOROS_BACKGROUND_JOB_CANCEL_METADATA
+_OUROBOROS_EXECUTION_SESSION_CANCEL_METADATA = (
+    _tool_specs._OUROBOROS_EXECUTION_SESSION_CANCEL_METADATA
+)
+_OUROBOROS_BACKGROUND_JOB_CANCEL_CONTROL_METADATA = (
+    _tool_specs._OUROBOROS_BACKGROUND_JOB_CANCEL_CONTROL_METADATA
+)
+_OUROBOROS_EXECUTION_SESSION_CANCEL_CONTROL_METADATA = (
+    _tool_specs._OUROBOROS_EXECUTION_SESSION_CANCEL_CONTROL_METADATA
+)
+_OUROBOROS_WORKSPACE_WRITE_TOOLS = _tool_specs._OUROBOROS_WORKSPACE_WRITE_TOOLS
 
-_OUROBOROS_BACKGROUND_TOOLS = frozenset(
-    {
-        "ouroboros_start_auto",
-        "ouroboros_start_execute_seed",
-        "ouroboros_start_evaluate",
-        "ouroboros_start_evolve_step",
-        "ouroboros_ralph",
-        "ouroboros_start_ralph",
-    }
+_interview_schemas = importlib.import_module(f"{__name__}.interview_schemas")
+_code_investigation_repo_inspection_tool_capabilities = (
+    _interview_schemas._code_investigation_repo_inspection_tool_capabilities
 )
-_OUROBOROS_STATUS_TOOLS = frozenset(
-    {
-        "ouroboros_session_status",
-        "ouroboros_job_status",
-        "ouroboros_job_wait",
-        "ouroboros_job_result",
-        "ouroboros_ac_tree_hud",
-        "ouroboros_query_events",
-        "ouroboros_query_projection",
-        "ouroboros_lineage_status",
-    }
+_interview_code_investigation_answer_contract = (
+    _interview_schemas._interview_code_investigation_answer_contract
 )
-_OUROBOROS_CANCEL_TOOLS = frozenset(
-    {
-        "ouroboros_cancel_job",
-        "ouroboros_cancel_execution",
-    }
+_interview_code_investigation_request_schema = (
+    _interview_schemas._interview_code_investigation_request_schema
 )
-_OUROBOROS_WORKSPACE_WRITE_TOOLS = frozenset(
-    {
-        "ouroboros_auto",
-        "ouroboros_start_auto",
-        "ouroboros_execute_seed",
-        "ouroboros_start_execute_seed",
-        "ouroboros_evolve_step",
-        "ouroboros_start_evolve_step",
-        "ouroboros_ralph",
-        "ouroboros_start_ralph",
-        "ouroboros_generate_seed",
-        "ouroboros_checklist_verify",
-        "ouroboros_brownfield",
-    }
+_interview_question_advisory_fanout_metadata = (
+    _interview_schemas._interview_question_advisory_fanout_metadata
 )
-_OUROBOROS_SUBAGENT_TOOLS = frozenset(
-    {
-        "ouroboros_interview",
-        "ouroboros_lateral_think",
-        "ouroboros_pm_interview",
-    }
+_interview_question_advisory_request_schema = (
+    _interview_schemas._interview_question_advisory_request_schema
 )
-_OUROBOROS_DEFAULT_EXECUTION_MODE = "blocking"
-_OUROBOROS_DEFAULT_RETRY_METADATA: Mapping[str, Any] = {
-    "supported": True,
-    "mode": "handler_owned",
-}
-_OUROBOROS_JOB_POLL_RETRY_METADATA: Mapping[str, Any] = {
-    "supported": True,
-    "mode": "job_poll",
-}
-_OUROBOROS_UNSUPPORTED_RETRY_METADATA: Mapping[str, Any] = {
-    "supported": False,
-    "mode": "unsupported",
-}
-_OUROBOROS_DEFAULT_INTERRUPT_METADATA: Mapping[str, Any] = {
-    "supported": True,
-    "mode": "soft",
-}
-_OUROBOROS_BLOCKING_INTERRUPT_METADATA: Mapping[str, Any] = {
-    "supported": True,
-    "mode": "soft",
-    "execution_mode": "blocking",
-    "blocking_semantics": "synchronous_handler",
-    "resumable": False,
-    "background_companions": (),
-    "target_context_keys": (),
-}
-_OUROBOROS_BACKGROUND_INTERRUPT_METADATA: Mapping[str, Any] = {
-    "supported": True,
-    "mode": "resumable_background_job",
-    "resumable": True,
-    "cancellable": True,
-    "resume_companions": (
-        "ouroboros_job_status",
-        "ouroboros_job_wait",
-        "ouroboros_job_result",
-    ),
-    "cancel_companions": ("ouroboros_cancel_job",),
-    "target_context_keys": ("job_id",),
-}
-_OUROBOROS_TERMINAL_CONTROL_INTERRUPT_METADATA_BY_TOOL: Mapping[str, Mapping[str, Any]] = {
-    "ouroboros_cancel_job": {
-        "supported": True,
-        "mode": "terminal_control",
-        "terminal_action": "cancel",
-        "target_type": "background_job",
-        "target_context_keys": ("job_id",),
-        "directive_semantics": "request_terminal_job_cancellation",
-        "terminal_statuses": ("cancelled",),
-        "idempotent": True,
-    },
-    "ouroboros_cancel_execution": {
-        "supported": True,
-        "mode": "terminal_control",
-        "terminal_action": "cancel",
-        "target_type": "execution_session",
-        "target_context_keys": ("execution_id",),
-        "directive_semantics": "request_terminal_execution_cancellation",
-        "terminal_statuses": ("cancelled",),
-        "idempotent": True,
-    },
-}
-_OUROBOROS_UNSUPPORTED_INTERRUPT_METADATA: Mapping[str, Any] = {
-    "supported": False,
-    "mode": "unsupported",
-}
-_OUROBOROS_READ_ONLY_INTERRUPT_METADATA: Mapping[str, Any] = {
-    "supported": True,
-    "mode": "read_only_non_mutating",
-    "mutation_semantics": "no_state_mutation",
-    "resumable": False,
-    "target_context_keys": (),
-}
-_OUROBOROS_UNSUPPORTED_CANCEL_METADATA: Mapping[str, Any] = {
-    "supported": False,
-    "mode": "unsupported",
-    "companions": (),
-    "target_context_keys": (),
-}
-_OUROBOROS_SIDE_EFFECT_FREE_METADATA: tuple[str, ...] = ()
-_OUROBOROS_MUTATION_TARGETS_BY_SIDE_EFFECT: Mapping[str, tuple[str, ...]] = {
-    "background_job_start": ("background_job",),
-    "checkpoint_store_write": ("checkpoint_store",),
-    "event_store_write": ("event_store",),
-    "runtime_control": ("runtime",),
-    "session_state_write": ("session_state",),
-    "side_effect_free": (),
-    "subagent_dispatch": ("subagent",),
-    "workspace_write": ("workspace",),
-}
-_OUROBOROS_STATE_MUTATIONS_BY_TOOL: Mapping[str, tuple[Mapping[str, Any], ...]] = {
-    "ouroboros_auto": (
-        {
-            "target": "auto_session_state",
-            "operation": "run_auto_pipeline_to_seed_and_execution",
-            "side_effect": "event_store_write",
-            "context_keys": ("goal", "resume", "cwd"),
-        },
-        {
-            "target": "workspace",
-            "operation": "apply_auto_generated_execution_changes",
-            "side_effect": "workspace_write",
-            "context_keys": ("cwd",),
-        },
-    ),
-    "ouroboros_start_auto": (
-        {
-            "target": "auto_session_state",
-            "operation": "enqueue_background_auto_pipeline",
-            "side_effect": "event_store_write",
-            "context_keys": ("goal", "resume", "cwd"),
-        },
-        {
-            "target": "workspace",
-            "operation": "apply_background_auto_execution_changes",
-            "side_effect": "workspace_write",
-            "context_keys": ("cwd",),
-        },
-    ),
-    "ouroboros_execute_seed": (
-        {
-            "target": "execution_session",
-            "operation": "run_seed_execution_session",
-            "side_effect": "event_store_write",
-            "context_keys": ("seed_path", "seed_content", "session_id", "cwd"),
-        },
-        {
-            "target": "workspace",
-            "operation": "apply_seed_execution_changes",
-            "side_effect": "workspace_write",
-            "context_keys": ("cwd",),
-        },
-    ),
-    "ouroboros_start_execute_seed": (
-        {
-            "target": "execution_session",
-            "operation": "enqueue_background_seed_execution",
-            "side_effect": "event_store_write",
-            "context_keys": ("seed_content", "session_id", "cwd"),
-        },
-        {
-            "target": "workspace",
-            "operation": "apply_background_seed_execution_changes",
-            "side_effect": "workspace_write",
-            "context_keys": ("cwd",),
-        },
-    ),
-    "ouroboros_evaluate": (
-        {
-            "target": "session_state",
-            "operation": "append_evaluation_result",
-            "side_effect": "session_state_write",
-            "context_keys": ("session_id",),
-        },
-    ),
-    "ouroboros_evolve_step": (
-        {
-            "target": "lineage_state",
-            "operation": "append_evolution_generation_result",
-            "side_effect": "event_store_write",
-            "context_keys": ("lineage_id", "seed_content", "execution_id"),
-        },
-        {
-            "target": "workspace",
-            "operation": "apply_evolution_generation_changes",
-            "side_effect": "workspace_write",
-            "context_keys": ("project_dir",),
-        },
-    ),
-    "ouroboros_evolve_rewind": (
-        {
-            "target": "lineage_state",
-            "operation": "truncate_generations_after_target",
-            "side_effect": "session_state_write",
-            "context_keys": ("lineage_id", "to_generation"),
-        },
-    ),
-    "ouroboros_start_evolve_step": (
-        {
-            "target": "lineage_state",
-            "operation": "enqueue_background_evolution_generation",
-            "side_effect": "event_store_write",
-            "context_keys": ("lineage_id", "seed_content", "execution_id"),
-        },
-        {
-            "target": "workspace",
-            "operation": "apply_background_evolution_generation_changes",
-            "side_effect": "workspace_write",
-            "context_keys": ("project_dir",),
-        },
-    ),
-    "ouroboros_interview": (
-        {
-            "target": "interview_state",
-            "operation": "create_or_update_or_complete_interview_session",
-            "side_effect": "session_state_write",
-            "context_keys": (
-                "initial_context",
-                "session_id",
-                "answer",
-                "last_question",
-            ),
-        },
-        {
-            "target": "subagent_dispatch_log",
-            "operation": "record_interview_subagent_dispatch",
-            "side_effect": "subagent_dispatch",
-            "context_keys": ("session_id",),
-        },
-    ),
-    "ouroboros_lateral_think": (
-        {
-            "target": "lateral_panel_state",
-            "operation": "dispatch_persona_panel_and_synthesize_findings",
-            "side_effect": "subagent_dispatch",
-            "context_keys": (
-                "problem_context",
-                "current_approach",
-                "failed_attempts",
-            ),
-        },
-        {
-            "target": "session_state",
-            "operation": "record_lateral_review_result",
-            "side_effect": "session_state_write",
-            "context_keys": ("problem_context",),
-        },
-    ),
-    "ouroboros_measure_drift": (
-        {
-            "target": "session_state",
-            "operation": "append_drift_measurement",
-            "side_effect": "session_state_write",
-            "context_keys": ("session_id",),
-        },
-    ),
-    "ouroboros_pm_interview": (
-        {
-            "target": "interview_state",
-            "operation": "create_or_update_pm_interview_session",
-            "side_effect": "session_state_write",
-            "context_keys": (
-                "initial_context",
-                "session_id",
-                "answer",
-                "last_question",
-            ),
-        },
-        {
-            "target": "pm_meta_state",
-            "operation": "persist_pm_session_metadata",
-            "side_effect": "session_state_write",
-            "context_keys": ("session_id", "cwd", "selected_repos"),
-        },
-        {
-            "target": "subagent_dispatch_log",
-            "operation": "record_pm_interview_subagent_dispatch",
-            "side_effect": "subagent_dispatch",
-            "context_keys": ("session_id",),
-        },
-    ),
-    "ouroboros_qa": (
-        {
-            "target": "qa_session_state",
-            "operation": "append_qa_iteration_verdict",
-            "side_effect": "session_state_write",
-            "context_keys": ("qa_session_id", "iteration_history"),
-        },
-    ),
-    "ouroboros_ralph": (
-        {
-            "target": "ralph_loop_state",
-            "operation": "run_evolution_loop_until_terminal_condition",
-            "side_effect": "event_store_write",
-            "context_keys": ("lineage_id", "execution_id"),
-        },
-        {
-            "target": "workspace",
-            "operation": "apply_ralph_loop_generation_changes",
-            "side_effect": "workspace_write",
-            "context_keys": ("project_dir",),
-        },
-    ),
-    "ouroboros_start_ralph": (
-        {
-            "target": "ralph_loop_state",
-            "operation": "enqueue_background_evolution_loop",
-            "side_effect": "event_store_write",
-            "context_keys": ("lineage_id", "execution_id"),
-        },
-        {
-            "target": "workspace",
-            "operation": "apply_background_ralph_loop_generation_changes",
-            "side_effect": "workspace_write",
-            "context_keys": ("project_dir",),
-        },
-    ),
-    "ouroboros_cancel_execution": (
-        {
-            "target": "execution_session",
-            "operation": "mark_execution_session_cancelled",
-            "side_effect": "session_state_write",
-            "context_keys": ("execution_id", "reason"),
-        },
-        {
-            "target": "event_store",
-            "operation": "append_session_cancelled_event",
-            "side_effect": "event_store_write",
-            "context_keys": ("execution_id", "reason"),
-        },
-        {
-            "target": "runtime",
-            "operation": "signal_execution_runner_to_stop_via_cancellation_event",
-            "side_effect": "runtime_control",
-            "context_keys": ("execution_id",),
-        },
-    ),
-    "ouroboros_cancel_job": (
-        {
-            "target": "background_job",
-            "operation": "mark_background_job_cancel_requested",
-            "side_effect": "event_store_write",
-            "context_keys": ("job_id",),
-        },
-        {
-            "target": "checkpoint_store",
-            "operation": "persist_durable_agent_process_cancel_signal",
-            "side_effect": "checkpoint_store_write",
-            "context_keys": ("job_id",),
-        },
-        {
-            "target": "runtime",
-            "operation": "cancel_live_background_job_tasks",
-            "side_effect": "runtime_control",
-            "context_keys": ("job_id",),
-        },
-        {
-            "target": "session_state",
-            "operation": "mark_linked_execution_session_cancelled_when_needed",
-            "side_effect": "session_state_write",
-            "context_keys": ("job_id",),
-        },
-    ),
-}
-_OUROBOROS_BACKGROUND_JOB_CANCEL_METADATA: Mapping[str, Any] = {
-    "supported": True,
-    "mode": "background_job",
-    "companions": ("ouroboros_cancel_job",),
-    "target_context_keys": ("job_id",),
-}
-_OUROBOROS_EXECUTION_SESSION_CANCEL_METADATA: Mapping[str, Any] = {
-    "supported": True,
-    "mode": "execution_session",
-    "companions": ("ouroboros_cancel_execution",),
-    "target_context_keys": ("execution_id",),
-}
-_OUROBOROS_BACKGROUND_JOB_CANCEL_CONTROL_METADATA: Mapping[str, Any] = {
-    "supported": True,
-    "mode": "background_job_control",
-    "companions": (
-        "ouroboros_job_status",
-        "ouroboros_job_wait",
-        "ouroboros_job_result",
-    ),
-    "target_context_keys": ("job_id",),
-}
-_OUROBOROS_EXECUTION_SESSION_CANCEL_CONTROL_METADATA: Mapping[str, Any] = {
-    "supported": True,
-    "mode": "execution_session_control",
-    "companions": (
-        "ouroboros_execute_seed",
-        "ouroboros_start_execute_seed",
-    ),
-    "target_context_keys": ("execution_id",),
-}
-
-_OUROBOROS_TOOL_CAPABILITY_SPECS: Mapping[str, _OuroborosToolCapabilitySpec] = {
-    "ouroboros_execute_seed": _OuroborosToolCapabilitySpec(
-        execution_mode="blocking",
-        companions=(
-            "ouroboros_start_execute_seed",
-            "ouroboros_cancel_execution",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-        ),
-        side_effects=("workspace_write", "event_store_write"),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_BLOCKING_INTERRUPT_METADATA,
-    ),
-    "ouroboros_start_execute_seed": _OuroborosToolCapabilitySpec(
-        execution_mode="background",
-        companions=(
-            "ouroboros_execute_seed",
-            "ouroboros_cancel_execution",
-            "ouroboros_start_evaluate",
-            "ouroboros_start_evolve_step",
-            "ouroboros_start_ralph",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-        ),
-        side_effects=("workspace_write", "event_store_write"),
-        retry=_OUROBOROS_JOB_POLL_RETRY_METADATA,
-        interrupt=_OUROBOROS_BACKGROUND_INTERRUPT_METADATA,
-    ),
-    "ouroboros_auto": _OuroborosToolCapabilitySpec(
-        execution_mode="blocking",
-        companions=("ouroboros_start_auto",),
-        side_effects=("workspace_write", "event_store_write"),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_BLOCKING_INTERRUPT_METADATA,
-    ),
-    "ouroboros_start_auto": _OuroborosToolCapabilitySpec(
-        execution_mode="background",
-        companions=(
-            "ouroboros_auto",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-        ),
-        side_effects=("workspace_write", "event_store_write"),
-        retry=_OUROBOROS_JOB_POLL_RETRY_METADATA,
-        interrupt=_OUROBOROS_BACKGROUND_INTERRUPT_METADATA,
-    ),
-    "ouroboros_session_status": _OuroborosToolCapabilitySpec(
-        execution_mode="status",
-        companions=(
-            "ouroboros_query_events",
-            "ouroboros_query_projection",
-            "ouroboros_ac_tree_hud",
-        ),
-        side_effects=_OUROBOROS_SIDE_EFFECT_FREE_METADATA,
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_READ_ONLY_INTERRUPT_METADATA,
-        mutation_class=CapabilityMutationClass.READ_ONLY,
-    ),
-    "ouroboros_job_status": _OuroborosToolCapabilitySpec(
-        execution_mode="status",
-        companions=(
-            "ouroboros_start_auto",
-            "ouroboros_start_execute_seed",
-            "ouroboros_start_evaluate",
-            "ouroboros_start_evolve_step",
-            "ouroboros_start_ralph",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-        ),
-        side_effects=_OUROBOROS_SIDE_EFFECT_FREE_METADATA,
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_READ_ONLY_INTERRUPT_METADATA,
-        mutation_class=CapabilityMutationClass.READ_ONLY,
-    ),
-    "ouroboros_job_wait": _OuroborosToolCapabilitySpec(
-        execution_mode="status",
-        companions=(
-            "ouroboros_start_auto",
-            "ouroboros_start_execute_seed",
-            "ouroboros_start_evaluate",
-            "ouroboros_start_evolve_step",
-            "ouroboros_start_ralph",
-            "ouroboros_job_status",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-        ),
-        side_effects=_OUROBOROS_SIDE_EFFECT_FREE_METADATA,
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_READ_ONLY_INTERRUPT_METADATA,
-        mutation_class=CapabilityMutationClass.READ_ONLY,
-    ),
-    "ouroboros_job_result": _OuroborosToolCapabilitySpec(
-        execution_mode="status",
-        companions=(
-            "ouroboros_start_auto",
-            "ouroboros_start_execute_seed",
-            "ouroboros_start_evaluate",
-            "ouroboros_start_evolve_step",
-            "ouroboros_start_ralph",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_cancel_job",
-        ),
-        side_effects=_OUROBOROS_SIDE_EFFECT_FREE_METADATA,
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_READ_ONLY_INTERRUPT_METADATA,
-        mutation_class=CapabilityMutationClass.READ_ONLY,
-    ),
-    "ouroboros_ac_tree_hud": _OuroborosToolCapabilitySpec(
-        execution_mode="status",
-        companions=(
-            "ouroboros_session_status",
-            "ouroboros_query_events",
-            "ouroboros_query_projection",
-        ),
-        side_effects=_OUROBOROS_SIDE_EFFECT_FREE_METADATA,
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_READ_ONLY_INTERRUPT_METADATA,
-        mutation_class=CapabilityMutationClass.READ_ONLY,
-    ),
-    "ouroboros_cancel_job": _OuroborosToolCapabilitySpec(
-        execution_mode="cancel",
-        companions=(
-            "ouroboros_start_auto",
-            "ouroboros_start_execute_seed",
-            "ouroboros_start_evaluate",
-            "ouroboros_start_evolve_step",
-            "ouroboros_start_ralph",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-        ),
-        side_effects=(
-            "runtime_control",
-            "event_store_write",
-            "checkpoint_store_write",
-            "session_state_write",
-        ),
-        retry=_OUROBOROS_UNSUPPORTED_RETRY_METADATA,
-        interrupt=_OUROBOROS_TERMINAL_CONTROL_INTERRUPT_METADATA_BY_TOOL["ouroboros_cancel_job"],
-    ),
-    "ouroboros_query_events": _OuroborosToolCapabilitySpec(
-        execution_mode="status",
-        companions=(
-            "ouroboros_session_status",
-            "ouroboros_query_projection",
-            "ouroboros_ac_tree_hud",
-        ),
-        side_effects=_OUROBOROS_SIDE_EFFECT_FREE_METADATA,
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_READ_ONLY_INTERRUPT_METADATA,
-        mutation_class=CapabilityMutationClass.READ_ONLY,
-    ),
-    "ouroboros_query_projection": _OuroborosToolCapabilitySpec(
-        execution_mode="status",
-        companions=(
-            "ouroboros_session_status",
-            "ouroboros_query_events",
-            "ouroboros_ac_tree_hud",
-        ),
-        side_effects=_OUROBOROS_SIDE_EFFECT_FREE_METADATA,
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_READ_ONLY_INTERRUPT_METADATA,
-        mutation_class=CapabilityMutationClass.READ_ONLY,
-    ),
-    "ouroboros_generate_seed": _OuroborosToolCapabilitySpec(
-        execution_mode="blocking",
-        companions=(
-            "ouroboros_interview",
-            "ouroboros_pm_interview",
-            "ouroboros_brownfield",
-        ),
-        side_effects=("workspace_write", "event_store_write"),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_BLOCKING_INTERRUPT_METADATA,
-    ),
-    "ouroboros_measure_drift": _OuroborosToolCapabilitySpec(
-        execution_mode="blocking",
-        companions=(
-            "ouroboros_evaluate",
-            "ouroboros_start_evaluate",
-            "ouroboros_checklist_verify",
-            "ouroboros_qa",
-        ),
-        side_effects=("session_state_write",),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_BLOCKING_INTERRUPT_METADATA,
-    ),
-    "ouroboros_interview": _OuroborosToolCapabilitySpec(
-        execution_mode="subagent_orchestration",
-        companions=(
-            "ouroboros_generate_seed",
-            "ouroboros_pm_interview",
-            "ouroboros_brownfield",
-            "ouroboros_lateral_think",
-        ),
-        side_effects=("subagent_dispatch", "session_state_write"),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_DEFAULT_INTERRUPT_METADATA,
-    ),
-    "ouroboros_evaluate": _OuroborosToolCapabilitySpec(
-        execution_mode="blocking",
-        companions=(
-            "ouroboros_start_evaluate",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-            "ouroboros_checklist_verify",
-            "ouroboros_measure_drift",
-            "ouroboros_qa",
-        ),
-        side_effects=("session_state_write",),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_BLOCKING_INTERRUPT_METADATA,
-    ),
-    "ouroboros_start_evaluate": _OuroborosToolCapabilitySpec(
-        execution_mode="background",
-        companions=(
-            "ouroboros_start_execute_seed",
-            "ouroboros_start_evolve_step",
-            "ouroboros_start_ralph",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-            "ouroboros_evaluate",
-            "ouroboros_checklist_verify",
-            "ouroboros_measure_drift",
-            "ouroboros_qa",
-        ),
-        side_effects=("background_job_start", "event_store_write"),
-        retry=_OUROBOROS_JOB_POLL_RETRY_METADATA,
-        interrupt=_OUROBOROS_BACKGROUND_INTERRUPT_METADATA,
-    ),
-    "ouroboros_checklist_verify": _OuroborosToolCapabilitySpec(
-        execution_mode="blocking",
-        companions=(
-            "ouroboros_evaluate",
-            "ouroboros_start_evaluate",
-            "ouroboros_measure_drift",
-            "ouroboros_qa",
-        ),
-        side_effects=("workspace_write", "event_store_write"),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_BLOCKING_INTERRUPT_METADATA,
-    ),
-    "ouroboros_lateral_think": _OuroborosToolCapabilitySpec(
-        execution_mode="subagent_orchestration",
-        companions=("ouroboros_interview",),
-        side_effects=("subagent_dispatch", "session_state_write"),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_DEFAULT_INTERRUPT_METADATA,
-    ),
-    "ouroboros_evolve_step": _OuroborosToolCapabilitySpec(
-        execution_mode="blocking",
-        companions=(
-            "ouroboros_start_evolve_step",
-            "ouroboros_lineage_status",
-            "ouroboros_evolve_rewind",
-            "ouroboros_ralph",
-            "ouroboros_start_ralph",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-        ),
-        side_effects=("workspace_write", "event_store_write"),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_BLOCKING_INTERRUPT_METADATA,
-    ),
-    "ouroboros_start_evolve_step": _OuroborosToolCapabilitySpec(
-        execution_mode="background",
-        companions=(
-            "ouroboros_start_execute_seed",
-            "ouroboros_start_evaluate",
-            "ouroboros_start_ralph",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-            "ouroboros_evolve_step",
-            "ouroboros_lineage_status",
-            "ouroboros_evolve_rewind",
-            "ouroboros_ralph",
-        ),
-        side_effects=("workspace_write", "event_store_write"),
-        retry=_OUROBOROS_JOB_POLL_RETRY_METADATA,
-        interrupt=_OUROBOROS_BACKGROUND_INTERRUPT_METADATA,
-    ),
-    "ouroboros_ralph": _OuroborosToolCapabilitySpec(
-        execution_mode="background",
-        companions=(
-            "ouroboros_evolve_step",
-            "ouroboros_start_evolve_step",
-            "ouroboros_lineage_status",
-            "ouroboros_evolve_rewind",
-            "ouroboros_start_ralph",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-        ),
-        side_effects=("workspace_write", "event_store_write"),
-        retry=_OUROBOROS_JOB_POLL_RETRY_METADATA,
-        interrupt=_OUROBOROS_BACKGROUND_INTERRUPT_METADATA,
-    ),
-    "ouroboros_start_ralph": _OuroborosToolCapabilitySpec(
-        execution_mode="background",
-        companions=(
-            "ouroboros_start_execute_seed",
-            "ouroboros_start_evaluate",
-            "ouroboros_start_evolve_step",
-            "ouroboros_job_status",
-            "ouroboros_job_wait",
-            "ouroboros_job_result",
-            "ouroboros_cancel_job",
-            "ouroboros_evolve_step",
-            "ouroboros_lineage_status",
-            "ouroboros_evolve_rewind",
-            "ouroboros_ralph",
-        ),
-        side_effects=("workspace_write", "event_store_write"),
-        retry=_OUROBOROS_JOB_POLL_RETRY_METADATA,
-        interrupt=_OUROBOROS_BACKGROUND_INTERRUPT_METADATA,
-    ),
-    "ouroboros_lineage_status": _OuroborosToolCapabilitySpec(
-        execution_mode="status",
-        companions=(
-            "ouroboros_evolve_step",
-            "ouroboros_start_evolve_step",
-            "ouroboros_evolve_rewind",
-            "ouroboros_ralph",
-            "ouroboros_start_ralph",
-        ),
-        side_effects=_OUROBOROS_SIDE_EFFECT_FREE_METADATA,
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_READ_ONLY_INTERRUPT_METADATA,
-        mutation_class=CapabilityMutationClass.READ_ONLY,
-    ),
-    "ouroboros_evolve_rewind": _OuroborosToolCapabilitySpec(
-        execution_mode="blocking",
-        companions=(
-            "ouroboros_evolve_step",
-            "ouroboros_start_evolve_step",
-            "ouroboros_lineage_status",
-            "ouroboros_ralph",
-            "ouroboros_start_ralph",
-        ),
-        side_effects=("session_state_write",),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_BLOCKING_INTERRUPT_METADATA,
-    ),
-    "ouroboros_cancel_execution": _OuroborosToolCapabilitySpec(
-        execution_mode="cancel",
-        companions=("ouroboros_execute_seed", "ouroboros_start_execute_seed"),
-        side_effects=(
-            "runtime_control",
-            "event_store_write",
-            "session_state_write",
-        ),
-        retry=_OUROBOROS_UNSUPPORTED_RETRY_METADATA,
-        interrupt=_OUROBOROS_TERMINAL_CONTROL_INTERRUPT_METADATA_BY_TOOL[
-            "ouroboros_cancel_execution"
-        ],
-    ),
-    "ouroboros_brownfield": _OuroborosToolCapabilitySpec(
-        execution_mode="blocking",
-        companions=(
-            "ouroboros_interview",
-            "ouroboros_generate_seed",
-            "ouroboros_pm_interview",
-        ),
-        side_effects=("workspace_write", "event_store_write"),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_BLOCKING_INTERRUPT_METADATA,
-    ),
-    "ouroboros_pm_interview": _OuroborosToolCapabilitySpec(
-        execution_mode="subagent_orchestration",
-        companions=(
-            "ouroboros_interview",
-            "ouroboros_generate_seed",
-            "ouroboros_brownfield",
-        ),
-        side_effects=("subagent_dispatch", "session_state_write"),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_DEFAULT_INTERRUPT_METADATA,
-    ),
-    "ouroboros_qa": _OuroborosToolCapabilitySpec(
-        execution_mode="blocking",
-        companions=(
-            "ouroboros_evaluate",
-            "ouroboros_start_evaluate",
-            "ouroboros_checklist_verify",
-            "ouroboros_measure_drift",
-        ),
-        side_effects=("session_state_write",),
-        retry=_OUROBOROS_DEFAULT_RETRY_METADATA,
-        interrupt=_OUROBOROS_BLOCKING_INTERRUPT_METADATA,
-    ),
-}
-
-_OUROBOROS_CANCEL_METADATA: Mapping[str, Mapping[str, Any]] = {
-    "ouroboros_ac_tree_hud": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_auto": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_brownfield": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_cancel_execution": _OUROBOROS_EXECUTION_SESSION_CANCEL_CONTROL_METADATA,
-    "ouroboros_cancel_job": _OUROBOROS_BACKGROUND_JOB_CANCEL_CONTROL_METADATA,
-    "ouroboros_checklist_verify": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_evaluate": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_evolve_rewind": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_evolve_step": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_execute_seed": _OUROBOROS_EXECUTION_SESSION_CANCEL_METADATA,
-    "ouroboros_generate_seed": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_interview": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_job_result": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_job_status": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_job_wait": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_lateral_think": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_lineage_status": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_measure_drift": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_pm_interview": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_qa": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_query_events": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_query_projection": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_ralph": _OUROBOROS_BACKGROUND_JOB_CANCEL_METADATA,
-    "ouroboros_session_status": _OUROBOROS_UNSUPPORTED_CANCEL_METADATA,
-    "ouroboros_start_auto": _OUROBOROS_BACKGROUND_JOB_CANCEL_METADATA,
-    "ouroboros_start_evaluate": _OUROBOROS_BACKGROUND_JOB_CANCEL_METADATA,
-    "ouroboros_start_evolve_step": _OUROBOROS_BACKGROUND_JOB_CANCEL_METADATA,
-    "ouroboros_start_execute_seed": _OUROBOROS_BACKGROUND_JOB_CANCEL_METADATA,
-    "ouroboros_start_ralph": _OUROBOROS_BACKGROUND_JOB_CANCEL_METADATA,
-}
-
-_OUROBOROS_BACKGROUND_BLOCKING_COMPANIONS: Mapping[str, str] = {
-    "ouroboros_start_auto": "ouroboros_auto",
-    "ouroboros_start_evaluate": "ouroboros_evaluate",
-    "ouroboros_start_evolve_step": "ouroboros_evolve_step",
-    "ouroboros_start_execute_seed": "ouroboros_execute_seed",
-}
-
-_OUROBOROS_BACKGROUND_LIFECYCLE_ROLE_TOOLS: Mapping[str, str] = {
-    "status": "ouroboros_job_status",
-    "wait": "ouroboros_job_wait",
-    "result": "ouroboros_job_result",
-    "cancel": "ouroboros_cancel_job",
-}
-_OUROBOROS_JOB_LIFECYCLE_SIBLING_ORDER = (
-    "status",
-    "wait",
-    "result",
-    "cancel",
+interview_code_investigation_answer_contract = (
+    _interview_schemas.interview_code_investigation_answer_contract
 )
 
 
@@ -1488,734 +535,6 @@ def _cancel_metadata_for_ouroboros_tool(name: str) -> Mapping[str, Any]:
         if isinstance(value, tuple)
         else value
         for key, value in metadata.items()
-    }
-
-
-def _interview_code_investigation_request_schema() -> dict[str, Any]:
-    """Return the runtime request model for interview code-fact investigation."""
-    target_schema: dict[str, Any] = {
-        "type": "object",
-        "oneOf": [
-            {
-                "title": "WorkspaceTarget",
-                "additionalProperties": False,
-                "required": ["target_type", "scope"],
-                "properties": {
-                    "target_type": {"const": "workspace"},
-                    "scope": {
-                        "type": "string",
-                        "enum": ["active", "selected_repositories", "all_available"],
-                    },
-                },
-            },
-            {
-                "title": "RelativePathTarget",
-                "additionalProperties": False,
-                "required": ["target_type", "path"],
-                "properties": {
-                    "target_type": {"const": "relative_path"},
-                    "path": {
-                        "type": "string",
-                        "minLength": 1,
-                        "description": "Repository-relative file or directory path.",
-                    },
-                },
-            },
-            {
-                "title": "GlobTarget",
-                "additionalProperties": False,
-                "required": ["target_type", "pattern"],
-                "properties": {
-                    "target_type": {"const": "glob"},
-                    "pattern": {
-                        "type": "string",
-                        "minLength": 1,
-                        "description": "Repository-relative glob pattern.",
-                    },
-                },
-            },
-            {
-                "title": "SymbolTarget",
-                "additionalProperties": False,
-                "required": ["target_type", "name"],
-                "properties": {
-                    "target_type": {"const": "symbol"},
-                    "name": {
-                        "type": "string",
-                        "minLength": 1,
-                        "description": "Function, class, module, command, or config symbol to locate.",
-                    },
-                    "path_hint": {
-                        "type": "string",
-                        "minLength": 1,
-                        "description": "Optional repository-relative search hint.",
-                    },
-                },
-            },
-        ],
-    }
-    return {
-        "type": "object",
-        "additionalProperties": False,
-        "required": [
-            "session_id",
-            "question_identity",
-            "question",
-            "investigation_goal",
-            "investigation_targets",
-            "fact_categories",
-            "allowed_capabilities",
-            "repo_inspection_tool_capabilities",
-            "confidence_policy",
-            "answer_prefixes",
-            "answer_contract",
-            "mcp_tool_capability",
-        ],
-        "properties": {
-            "session_id": {
-                "type": "string",
-                "description": "Current Ouroboros interview session ID.",
-            },
-            "question_identity": {
-                "type": "string",
-                "pattern": r"^interview-question:[0-9a-f]{16}$",
-                "description": (
-                    "Stable identity derived from the originating interview "
-                    "question using stable_code_investigation_question_identity()."
-                ),
-            },
-            "question": {
-                "type": "string",
-                "description": "The MCP-generated interview question requiring code facts.",
-            },
-            "last_question": {
-                "type": "string",
-                "description": "Previously asked question text, when available.",
-            },
-            "investigation_goal": {
-                "type": "string",
-                "enum": ["describe_current_state_from_code"],
-                "description": "Code investigation is descriptive only; decisions route to the user.",
-            },
-            "investigation_targets": {
-                "type": "array",
-                "minItems": 1,
-                "items": target_schema,
-                "description": "Repository-agnostic descriptors for the code facts to inspect.",
-            },
-            "fact_categories": {
-                "type": "array",
-                "minItems": 1,
-                "items": {
-                    "type": "string",
-                    "enum": [
-                        "tech_stack",
-                        "frameworks",
-                        "dependencies",
-                        "current_patterns",
-                        "architecture",
-                        "file_structure",
-                        "configuration",
-                    ],
-                },
-            },
-            "allowed_capabilities": {
-                "type": "array",
-                "minItems": 1,
-                "items": {"type": "string", "enum": ["inspect_code"]},
-                "description": "Runtime capability used for local code facts.",
-            },
-            "repo_inspection_tool_capabilities": {
-                "type": "array",
-                "minItems": 1,
-                "items": {
-                    "type": "object",
-                    "additionalProperties": True,
-                    "required": [
-                        "tool_name",
-                        "stable_id",
-                        "source_kind",
-                        "source_name",
-                        "input_schema",
-                        "mutation_class",
-                        "parallel_safety",
-                        "interruptibility",
-                        "approval_class",
-                        "origin",
-                        "scope",
-                        "execution_mode",
-                        "logical_capability",
-                        "side_effects",
-                        "fallback_used",
-                    ],
-                    "properties": {
-                        "tool_name": {"type": "string", "enum": ["Read", "Glob", "Grep"]},
-                        "source_kind": {"const": "builtin"},
-                        "execution_mode": {"const": "repo_inspection"},
-                        "logical_capability": {"const": "inspect_code"},
-                        "fallback_used": {"const": False},
-                    },
-                },
-                "description": (
-                    "Concrete runtime repo-inspection tools a code-fact "
-                    "subagent can use to satisfy allowed_capabilities=inspect_code."
-                ),
-            },
-            "confidence_policy": {
-                "type": "object",
-                "additionalProperties": False,
-                "required": [
-                    "auto_confirm_when",
-                    "confirmation_required_when",
-                    "human_judgment_when",
-                ],
-                "properties": {
-                    "auto_confirm_when": {
-                        "type": "array",
-                        "items": {"type": "string"},
-                    },
-                    "confirmation_required_when": {
-                        "type": "array",
-                        "items": {"type": "string"},
-                    },
-                    "human_judgment_when": {
-                        "type": "array",
-                        "items": {"type": "string"},
-                    },
-                },
-            },
-            "answer_prefixes": {
-                "type": "array",
-                "minItems": 1,
-                "items": {
-                    "type": "string",
-                    "enum": ["[from-code]", "[from-code][auto-confirmed]"],
-                },
-            },
-            "answer_contract": {
-                "const": _interview_code_investigation_answer_contract(),
-                "description": "Exact response contract attached to this investigation request.",
-            },
-            "mcp_tool_capability": {
-                "type": "object",
-                "additionalProperties": True,
-                "required": [
-                    "tool_name",
-                    "stable_id",
-                    "source_kind",
-                    "source_name",
-                    "input_schema",
-                    "mutation_class",
-                    "execution_mode",
-                    "companions",
-                    "required_context_keys",
-                    "mutation_targets",
-                    "state_mutations",
-                    "side_effects",
-                    "retry",
-                    "interrupt",
-                    "cancel",
-                    "fallback_used",
-                    "orchestration",
-                ],
-                "properties": {
-                    "tool_name": {"const": "ouroboros_interview"},
-                    "fallback_used": {"const": False},
-                },
-                "description": (
-                    "Explicit Ouroboros-owned MCP capability metadata for the "
-                    "tool that emitted this investigation request."
-                ),
-            },
-        },
-    }
-
-
-def _interview_code_investigation_answer_contract() -> dict[str, Any]:
-    """Return the answer contract for one code-fact investigation request."""
-    answer_schema: dict[str, Any] = {
-        "type": "object",
-        "additionalProperties": False,
-        "required": [
-            "session_id",
-            "question_identity",
-            "answer_prefix",
-            "answer_text",
-            "confidence",
-            "evidence",
-            "requires_user_confirmation",
-        ],
-        "properties": {
-            "session_id": {
-                "type": "string",
-                "description": "Current Ouroboros interview session ID.",
-            },
-            "question_identity": {
-                "type": "string",
-                "pattern": r"^interview-question:[0-9a-f]{16}$",
-                "description": "Matches the originating code investigation request.",
-            },
-            "answer_prefix": {
-                "type": "string",
-                "enum": ["[from-code]", "[from-code][auto-confirmed]"],
-                "description": "Prefix to prepend when forwarding the answer to interview MCP.",
-            },
-            "answer_text": {
-                "type": "string",
-                "minLength": 1,
-                "description": "Concise descriptive fact answer without prescription.",
-            },
-            "confidence": {
-                "type": "string",
-                "enum": ["high_exact_match", "medium_inferred", "low_uncertain"],
-            },
-            "evidence": {
-                "type": "array",
-                "minItems": 1,
-                "items": {
-                    "type": "object",
-                    "additionalProperties": False,
-                    "required": ["source", "claim"],
-                    "properties": {
-                        "source": {
-                            "type": "string",
-                            "minLength": 1,
-                            "description": "Repository-relative file, symbol, or manifest source.",
-                        },
-                        "claim": {
-                            "type": "string",
-                            "minLength": 1,
-                            "description": "The factual claim supported by this evidence.",
-                        },
-                        "locator": {
-                            "type": "string",
-                            "minLength": 1,
-                            "description": "Optional line, key, dependency, or symbol locator.",
-                        },
-                    },
-                },
-            },
-            "requires_user_confirmation": {
-                "type": "boolean",
-                "description": "True when the answer must be confirmed before forwarding.",
-            },
-            "user_confirmation_prompt": {
-                "type": "string",
-                "minLength": 1,
-                "description": "Prompt text to show when confirmation is required.",
-            },
-        },
-        "allOf": [
-            {
-                "if": {
-                    "properties": {"answer_prefix": {"const": "[from-code][auto-confirmed]"}},
-                    "required": ["answer_prefix"],
-                },
-                "then": {
-                    "properties": {
-                        "confidence": {"const": "high_exact_match"},
-                        "requires_user_confirmation": {"const": False},
-                    }
-                },
-            },
-            {
-                "if": {
-                    "properties": {"requires_user_confirmation": {"const": True}},
-                    "required": ["requires_user_confirmation"],
-                },
-                "then": {"required": ["user_confirmation_prompt"]},
-            },
-            {
-                "if": {
-                    "properties": {"answer_prefix": {"const": "[from-code]"}},
-                    "required": ["answer_prefix"],
-                },
-                "then": {
-                    "properties": {"requires_user_confirmation": {"const": True}},
-                    "required": ["user_confirmation_prompt"],
-                },
-            },
-        ],
-    }
-    return {
-        "contract_id": "code_fact_investigation_answer.v1",
-        "scope": "single_code_fact_investigation_request",
-        "response_model_schema": answer_schema,
-        "prefix_semantics": {
-            "[from-code][auto-confirmed]": {
-                "confidence": "high_exact_match",
-                "requires_user_confirmation": False,
-                "forwarding": "send_to_mcp_immediately",
-            },
-            "[from-code]": {
-                "confidence": "medium_or_low",
-                "requires_user_confirmation": True,
-                "forwarding": "confirm_with_user_before_mcp",
-            },
-        },
-        "evidence_policy": {
-            "minimum_items": 1,
-            "source_format": "repository_relative_path_or_symbol",
-            "server_local_paths_allowed": False,
-        },
-        "runtime_instruction": (
-            "Produce exactly one structured answer payload for the originating "
-            "question_identity. Use [from-code][auto-confirmed] only for an "
-            "unambiguous manifest/config exact match; otherwise require user "
-            "confirmation and use [from-code] after confirmation."
-        ),
-    }
-
-
-def interview_code_investigation_answer_contract() -> dict[str, Any]:
-    """Return the public code-fact answer contract for generated requests."""
-    return _interview_code_investigation_answer_contract()
-
-
-def _code_investigation_repo_inspection_tool_capabilities() -> tuple[dict[str, Any], ...]:
-    """Return concrete repo-inspection tool capabilities for code-fact subagents."""
-    tool_schemas: Mapping[str, Mapping[str, Any]] = {
-        "Read": {
-            "type": "object",
-            "additionalProperties": True,
-            "required": ["file_path"],
-            "properties": {
-                "file_path": {
-                    "type": "string",
-                    "minLength": 1,
-                    "description": "Repository-local file path to inspect.",
-                },
-                "offset": {"type": "integer", "minimum": 1},
-                "limit": {"type": "integer", "minimum": 1},
-            },
-        },
-        "Glob": {
-            "type": "object",
-            "additionalProperties": True,
-            "required": ["pattern"],
-            "properties": {
-                "pattern": {
-                    "type": "string",
-                    "minLength": 1,
-                    "description": "Repository-local glob pattern to enumerate.",
-                },
-                "path": {
-                    "type": "string",
-                    "minLength": 1,
-                    "description": "Optional repository-local search root.",
-                },
-            },
-        },
-        "Grep": {
-            "type": "object",
-            "additionalProperties": True,
-            "required": ["pattern"],
-            "properties": {
-                "pattern": {
-                    "type": "string",
-                    "minLength": 1,
-                    "description": "Search pattern for repository-local evidence.",
-                },
-                "path": {
-                    "type": "string",
-                    "minLength": 1,
-                    "description": "Optional repository-local file or directory scope.",
-                },
-                "glob": {
-                    "type": "string",
-                    "minLength": 1,
-                    "description": "Optional file glob narrowing the search.",
-                },
-            },
-        },
-    }
-    capabilities: list[dict[str, Any]] = []
-    for tool_name in ("Read", "Glob", "Grep"):
-        semantics = _BUILTIN_SEMANTICS[tool_name]
-        capabilities.append(
-            {
-                "tool_name": tool_name,
-                "stable_id": f"builtin:{tool_name}",
-                "source_kind": "builtin",
-                "source_name": "built-in",
-                "input_schema": dict(tool_schemas[tool_name]),
-                "mutation_class": semantics.mutation_class.value,
-                "parallel_safety": semantics.parallel_safety.value,
-                "interruptibility": semantics.interruptibility.value,
-                "approval_class": semantics.approval_class.value,
-                "origin": semantics.origin.value,
-                "scope": semantics.scope.value,
-                "execution_mode": "repo_inspection",
-                "logical_capability": "inspect_code",
-                "side_effects": ["side_effect_free"],
-                "fallback_used": False,
-            }
-        )
-    return tuple(capabilities)
-
-
-def _interview_question_advisory_request_schema() -> dict[str, Any]:
-    """Return the runtime request model for per-question answer assistance."""
-    return {
-        "type": "object",
-        "additionalProperties": False,
-        "required": [
-            "session_id",
-            "question_identity",
-            "question",
-            "phase",
-            "user_question_first",
-            "advisory_goal",
-            "parallel_preference",
-            "sequential_fallback",
-            "allowed_capabilities",
-            "lanes",
-            "synthesis_contract",
-            "mcp_tool_capability",
-        ],
-        "properties": {
-            "session_id": {
-                "type": "string",
-                "description": "Current Ouroboros interview session ID.",
-            },
-            "question_identity": {
-                "type": "string",
-                "pattern": r"^interview-question:[0-9a-f]{16}$",
-                "description": (
-                    "Stable identity derived from the originating interview "
-                    "question using stable_code_investigation_question_identity()."
-                ),
-            },
-            "question": {
-                "type": "string",
-                "minLength": 1,
-                "description": "The already user-visible MCP interview question.",
-            },
-            "last_question": {
-                "type": "string",
-                "description": "Previously asked question text, when available.",
-            },
-            "phase": {
-                "type": "string",
-                "enum": ["start", "resume_pending", "answer"],
-            },
-            "ambiguity_score": {
-                "type": ["number", "null"],
-                "minimum": 0,
-                "maximum": 1,
-            },
-            "milestone": {
-                "type": ["string", "null"],
-                "enum": ["initial", "progress", "refined", "ready", None],
-            },
-            "user_question_first": {
-                "const": True,
-                "description": (
-                    "The parent runtime must surface the interview question before "
-                    "or while advisory fanout runs; advisory must never hide the "
-                    "question behind background research."
-                ),
-            },
-            "advisory_goal": {
-                "const": "help_human_answer_interview_question",
-                "description": (
-                    "Generate concise answer options, uncertainty notes, and a "
-                    "recommended draft without mutating interview state."
-                ),
-            },
-            "parallel_preference": {
-                "const": "parallel_when_runtime_supports_subagents",
-            },
-            "sequential_fallback": {
-                "type": "object",
-                "additionalProperties": False,
-                "required": ["supported", "mode", "trigger"],
-                "properties": {
-                    "supported": {"const": True},
-                    "mode": {"const": "sequential_advisory_lane_dispatch"},
-                    "trigger": {"const": "runtime_has_no_native_parallel_subagent_primitive"},
-                },
-            },
-            "allowed_capabilities": {
-                "type": "array",
-                "minItems": 1,
-                "items": {
-                    "type": "string",
-                    "enum": ["inspect_code", "web_research", "run_lateral_review"],
-                },
-            },
-            "lanes": {
-                "type": "array",
-                "minItems": 1,
-                "items": {
-                    "type": "object",
-                    "additionalProperties": False,
-                    "required": ["lane_id", "purpose", "capability", "required"],
-                    "properties": {
-                        "lane_id": {
-                            "type": "string",
-                            "enum": [
-                                "code_context",
-                                "web_context",
-                                "ambiguity_contrarian",
-                                "answer_simplifier",
-                                "architecture_implications",
-                            ],
-                        },
-                        "purpose": {"type": "string", "minLength": 1},
-                        "capability": {
-                            "type": "string",
-                            "enum": ["inspect_code", "web_research", "run_lateral_review"],
-                        },
-                        "persona": {
-                            "type": "string",
-                            "enum": ["researcher", "contrarian", "simplifier", "architect"],
-                        },
-                        "required": {"type": "boolean"},
-                    },
-                },
-            },
-            "code_investigation_request": {
-                "type": "object",
-                "additionalProperties": True,
-                "description": (
-                    "Optional code-fact request emitted alongside this advisory; "
-                    "reuse it for the code_context lane when present."
-                ),
-            },
-            "synthesis_contract": {
-                "type": "object",
-                "additionalProperties": False,
-                "required": [
-                    "output_shape",
-                    "max_options",
-                    "include_recommended_draft",
-                    "preserve_user_agency",
-                    "forward_to_mcp_only_after_user_or_auto_confirm",
-                ],
-                "properties": {
-                    "output_shape": {
-                        "const": "answer_advisory",
-                    },
-                    "max_options": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 5,
-                    },
-                    "include_recommended_draft": {"type": "boolean"},
-                    "preserve_user_agency": {"const": True},
-                    "forward_to_mcp_only_after_user_or_auto_confirm": {"const": True},
-                },
-            },
-            "mcp_tool_capability": {
-                "type": "object",
-                "additionalProperties": True,
-                "required": [
-                    "tool_name",
-                    "stable_id",
-                    "source_kind",
-                    "source_name",
-                    "input_schema",
-                    "mutation_class",
-                    "execution_mode",
-                    "companions",
-                    "required_context_keys",
-                    "mutation_targets",
-                    "state_mutations",
-                    "side_effects",
-                    "retry",
-                    "interrupt",
-                    "cancel",
-                    "fallback_used",
-                    "orchestration",
-                ],
-                "properties": {
-                    "tool_name": {"const": "ouroboros_interview"},
-                    "fallback_used": {"const": False},
-                },
-            },
-        },
-    }
-
-
-def _interview_question_advisory_fanout_metadata() -> dict[str, Any]:
-    """Return structured metadata for parent-session interview answer help."""
-    lanes = [
-        {
-            "lane_id": "code_context",
-            "purpose": "Find repo-local facts that may answer or constrain the question.",
-            "capability": "inspect_code",
-            "required": False,
-        },
-        {
-            "lane_id": "web_context",
-            "purpose": (
-                "Check current external facts only when the question depends on "
-                "third-party APIs, pricing, standards, security, or recent changes."
-            ),
-            "capability": "web_research",
-            "required": False,
-        },
-        {
-            "lane_id": "ambiguity_contrarian",
-            "purpose": "Name hidden assumptions, missing decisions, and risky vague words.",
-            "capability": "run_lateral_review",
-            "persona": "contrarian",
-            "required": True,
-        },
-        {
-            "lane_id": "answer_simplifier",
-            "purpose": "Turn the question into easy choices or a concise answer draft.",
-            "capability": "run_lateral_review",
-            "persona": "simplifier",
-            "required": True,
-        },
-        {
-            "lane_id": "architecture_implications",
-            "purpose": (
-                "Check whether the answer would change system shape, ownership, "
-                "interfaces, or rollout strategy."
-            ),
-            "capability": "run_lateral_review",
-            "persona": "architect",
-            "required": False,
-        },
-    ]
-    return {
-        "contract_id": "interview_question_advisory_fanout.v1",
-        "mcp_tool": "ouroboros_interview",
-        "companion_tool": "ouroboros_lateral_think",
-        "dispatch_timing": "after_question_is_visible_to_user",
-        "parallel_preference": "parallel_when_runtime_supports_subagents",
-        "sequential_fallback": {
-            "supported": True,
-            "mode": "sequential_advisory_lane_dispatch",
-            "trigger": "runtime_has_no_native_parallel_subagent_primitive",
-        },
-        "request_model_schema": _interview_question_advisory_request_schema(),
-        "lanes": lanes,
-        "synthesis_contract": {
-            "output_shape": "answer_advisory",
-            "max_options": 3,
-            "include_recommended_draft": True,
-            "preserve_user_agency": True,
-            "forward_to_mcp_only_after_user_or_auto_confirm": True,
-        },
-        "response_payload_refs": {
-            "plugin": "parent_runtime.ouroboros_dispatch.children",
-            "result_correlation_key": "lane_id",
-            "requires_prose_parsing": False,
-            "synthesis_owner": "parent_session",
-        },
-        "runtime_instruction": (
-            "Show the MCP interview question to the user first, then fan out "
-            "advisory lanes for code context, current web facts when needed, "
-            "ambiguity critique, simplification, and architecture implications. "
-            "Read child task results as they complete and synthesize them into "
-            "two or three answer options or one recommended draft. Do not forward advisory text to "
-            "ouroboros_interview until the user approves, edits, or explicitly "
-            "chooses auto-confirm."
-        ),
     }
 
 
@@ -2690,163 +1009,6 @@ def ouroboros_tool_capability_metadata(tool_name: str) -> dict[str, Any]:
         "source_kind": descriptor.source_kind,
         "source_name": descriptor.source_name,
         **metadata,
-    }
-
-
-def _lateral_persona_panel_request_schema() -> dict[str, Any]:
-    """Return the structured request model for lateral persona panels."""
-    return {
-        "type": "object",
-        "additionalProperties": False,
-        "required": ["problem_context", "current_approach"],
-        "properties": {
-            "problem_context": {
-                "type": "string",
-                "minLength": 1,
-                "description": "The stuck state or problem being reframed.",
-            },
-            "current_approach": {
-                "type": "string",
-                "minLength": 1,
-                "description": "What has already been tried.",
-            },
-            "persona": {
-                "type": "string",
-                "enum": [
-                    "hacker",
-                    "researcher",
-                    "simplifier",
-                    "architect",
-                    "contrarian",
-                    "all",
-                ],
-            },
-            "personas": {
-                "type": "array",
-                "items": {
-                    "type": "string",
-                    "enum": [
-                        "hacker",
-                        "researcher",
-                        "simplifier",
-                        "architect",
-                        "contrarian",
-                    ],
-                },
-            },
-            "failed_attempts": {
-                "type": "array",
-                "items": {"type": "string"},
-            },
-        },
-    }
-
-
-def _lateral_persona_panel_metadata() -> LateralPersonaPanelMetadata:
-    """Return the metadata contract for lateral persona panel dispatch."""
-    persona_roles = (
-        ("hacker", "Finds unconventional workarounds"),
-        ("researcher", "Seeks additional information"),
-        ("simplifier", "Reduces complexity"),
-        ("architect", "Restructures the approach"),
-        ("contrarian", "Challenges assumptions"),
-    )
-    return LateralPersonaPanelMetadata(
-        panel_id="lateral_persona_panel.v1",
-        mcp_tool="ouroboros_lateral_think",
-        dispatch_modes=("plugin", "sequential"),
-        legacy_dispatch_modes=("inline_fallback",),
-        parallel_preference="parallel_when_runtime_supports_subagents",
-        sequential_fallback={
-            "supported": True,
-            "mode": "sequential_persona_payload_dispatch",
-            "trigger": "runtime_has_no_native_parallel_subagent_primitive",
-        },
-        personas=tuple(
-            LateralPersonaMetadata(
-                persona_id=persona_id,
-                role=role,
-                prompt={
-                    "source": "build_lateral_multi_subagent",
-                    "payload_field": "payloads[].prompt",
-                    "context_field": "payloads[].context",
-                    "requires_prose_parsing": False,
-                },
-                response_payload_ref={
-                    "plugin": "MCPToolResult.meta._subagents[persona_id]",
-                    "inline_meta": "MCPToolResult.meta.payloads[persona_id]",
-                    "inline_content": (
-                        "content sentinel ouroboros-lateral-inline-dispatch-v1.payloads[persona_id]"
-                    ),
-                },
-            )
-            for persona_id, role in persona_roles
-        ),
-        request_model_schema=_lateral_persona_panel_request_schema(),
-        response_payload_refs={
-            "plugin": "MCPToolResult.meta._subagents",
-            "inline_meta": "MCPToolResult.meta.payloads",
-            "inline_content": ("content sentinel ouroboros-lateral-inline-dispatch-v1.payloads"),
-            "result_correlation_key": "context.persona",
-            "requires_prose_parsing": False,
-        },
-        runtime_instruction=(
-            "Call ouroboros_lateral_think first. If the response delegates via "
-            "_subagents, consume those payloads. If it returns sequential "
-            "payload metadata and the runtime has a native subagent primitive, "
-            "dispatch each structured payload by context.persona; otherwise "
-            "process those payloads sequentially. Treat inline_fallback as a "
-            "legacy alias for sequential."
-        ),
-    )
-
-
-def _pm_interview_subagent_metadata() -> dict[str, Any]:
-    """Return the metadata contract for PM interview subagent dispatch."""
-    return {
-        "directive": "run_pm_interview_subagent",
-        "mcp_tool": "ouroboros_pm_interview",
-        "dispatch_modes": ["plugin"],
-        "payload_builder": "build_pm_interview_subagent",
-        "request_model_schema": {
-            "type": "object",
-            "additionalProperties": True,
-            "properties": {
-                "session_id": {"type": "string"},
-                "action": {
-                    "type": "string",
-                    "enum": ["start", "answer", "resume", "generate", "select_repos"],
-                    "default": "start",
-                },
-                "initial_context": {"type": "string"},
-                "answer": {"type": "string"},
-                "cwd": {"type": "string"},
-                "selected_repos": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                },
-            },
-        },
-        "response_payload_refs": {
-            "plugin": "MCPToolResult.meta._subagent",
-            "content_json": "MCPToolResult.content[0].text._subagent",
-            "result_correlation_key": "context.session_id",
-            "requires_prose_parsing": False,
-        },
-        "subagent_context_keys": [
-            "session_id",
-            "action",
-            "initial_context",
-            "answer",
-            "cwd",
-            "selected_repos",
-        ],
-        "runtime_instruction": (
-            "Dispatch PM interview work through the `_subagent` payload produced "
-            "by build_pm_interview_subagent. Preserve session_id/action context "
-            "and consume the structured payload directly; do not infer PM "
-            "subagent routing from prose."
-        ),
     }
